pyproject.toml
[project]
name = "crypto-deltalake"
version = "0.1.0"
description = "Getting crypto data and store them using deltalake"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "binance-connector>=3.12.0",
    "deltalake>=0.25.5",
    "polars>=1.27.1",
    "python-dotenv>=1.1.0",
    "websockets>=15.0.1",
    "aiohttp>=3.8.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "seaborn>=0.13.2",
]
